chore(web): rename component
