+++ ts-packages/web/src/app/spaces/[id]/_components/space-contents.tsx
