+++ ts-packages/web/src/app/spaces/[id]/committee/page.client.tsx
@@ -3,7 +3,7 @@
 import React from 'react';
 import SpaceHeader from '../_components/space-header';
 import SpaceCouponProgress from '../_components/coupon-progress';
-import SpaceContents from '../_components/space_contents';
+import SpaceContents from '../_components/space-contents';
 import SpaceFiles from '../_components/space_files';
 import { useRouter } from 'next/navigation';
 import { useRedeemCode, useSpaceById } from '@/lib/api/ratel_api';
